add timeout test cases, refactored test classes to top level for use in timeout tests
