+++ src/test/java/org/soulwing/mail/transport/ErrorThrowingTransportTest.java
@@ -25,12 +25,6 @@ import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.fail;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileReader;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Condition;
@@ -41,13 +35,7 @@ import javax.mail.MessagingException;
 import javax.mail.Session;
 import javax.mail.Transport;
 import javax.mail.event.ConnectionEvent;
-import javax.mail.event.ConnectionListener;
-import javax.mail.event.TransportEvent;
-import javax.mail.event.TransportListener;
-import javax.mail.internet.MimeMessage;
-import javax.mail.internet.MimeMessage.RecipientType;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -77,12 +65,12 @@ public class ErrorThrowingTransportTest {
 
     transport.connect();
 
-    ConnectionEvent event = listener.awaitEvent();
+    ConnectionEvent event = listener.awaitEvent(LISTENER_WAIT, MAX_LISTENER_WAIT);
     assertThat(event, is(not(nullValue())));
     assertThat(event.getType(), is(equalTo(ConnectionEvent.OPENED)));
     listener.reset();
     transport.close();
-    event = listener.awaitEvent();
+    event = listener.awaitEvent(LISTENER_WAIT, MAX_LISTENER_WAIT);
     assertThat(event, is(not(nullValue())));
     assertThat(event.getType(), is(equalTo(ConnectionEvent.CLOSED)));
   }
@@ -118,88 +106,6 @@ public class ErrorThrowingTransportTest {
     final Properties properties = new Properties();
     properties.setProperty("mail.transport.protocol", "error");
     return properties;
-  }
-
-  private static abstract class AbstractEventListener<T> {
-
-    private final Lock lock = new ReentrantLock();
-    private final Condition readyCondition = lock.newCondition();
-
-    private T event;
-
-    void reset() {
-      event = null;
-    }
-
-    T awaitEvent() throws InterruptedException {
-      lock.lock();
-      try {
-        final long start = System.currentTimeMillis();
-        long now = start;
-        while (event == null && (now - start) < MAX_LISTENER_WAIT) {
-          readyCondition.await(LISTENER_WAIT, TimeUnit.MILLISECONDS);
-          now = System.currentTimeMillis();
-        }
-        return event;
-      }
-      finally {
-        lock.unlock();
-      }
-    }
-
-    void receiveEvent(T event) {
-      lock.lock();
-      try {
-        this.event = event;
-        readyCondition.signalAll();
-      }
-      finally {
-        lock.unlock();
-      }
-    }
-
-  }
-
-  private static class MockConnectionListener
-      extends AbstractEventListener<ConnectionEvent>
-      implements ConnectionListener {
-
-    @Override
-    public void opened(ConnectionEvent connectionEvent) {
-      receiveEvent(connectionEvent);
-    }
-
-    @Override
-    public void disconnected(ConnectionEvent connectionEvent) {
-      receiveEvent(connectionEvent);
-    }
-
-    @Override
-    public void closed(ConnectionEvent connectionEvent) {
-      receiveEvent(connectionEvent);
-    }
-
-  }
-
-  private static class MockTransportListener
-      extends AbstractEventListener<TransportEvent>
-      implements TransportListener {
-
-    @Override
-    public void messageDelivered(TransportEvent transportEvent) {
-      receiveEvent(transportEvent);
-    }
-
-    @Override
-    public void messageNotDelivered(TransportEvent transportEvent) {
-      receiveEvent(transportEvent);
-    }
-
-    @Override
-    public void messagePartiallyDelivered(TransportEvent transportEvent) {
-      receiveEvent(transportEvent);
-    }
-
   }
 
 }
